+++ server/component/assetDelete/tpl_success_rm_file.php
@@ -1,0 +1,7 @@
+<div class="container mt-3">
+    <div class="jumbotron">
+    <h1>Success</h1>
+        <p>The file was successfully removed from the server.</p>
+        <a href="<?php echo $url; ?>" class="btn btn-primary">Back to the Assets</a>
+    </div>
+</div>